Moved math classes in new math package
