+++ src/main/java/de/officeryoda/fhysics/engine/FhysicsCore.kt
@@ -1,5 +1,7 @@
 package de.officeryoda.fhysics.engine
 
+import de.officeryoda.fhysics.engine.math.BoundingBox
+import de.officeryoda.fhysics.engine.math.Vector2
 import de.officeryoda.fhysics.engine.objects.FhysicsObject
 import de.officeryoda.fhysics.extensions.times
 import de.officeryoda.fhysics.rendering.BetterSceneListener
