+++ src/main/java/de/officeryoda/fhysics/engine/collision/CollisionFinder.kt
@@ -1,8 +1,8 @@
 package de.officeryoda.fhysics.engine.collision
 
-import de.officeryoda.fhysics.engine.Projection
-import de.officeryoda.fhysics.engine.ProjectionResult
-import de.officeryoda.fhysics.engine.Vector2
+import de.officeryoda.fhysics.engine.math.Projection
+import de.officeryoda.fhysics.engine.math.ProjectionResult
+import de.officeryoda.fhysics.engine.math.Vector2
 import de.officeryoda.fhysics.engine.objects.*
 import kotlin.math.abs
 import kotlin.math.min
