+++ src/main/java/de/officeryoda/fhysics/engine/math/Matrix2x3.kt
@@ -1,4 +1,4 @@
-package de.officeryoda.fhysics.engine
+package de.officeryoda.fhysics.engine.math
 
 /**
  * This class is used for the transformation of Vector2 objects.
