+++ src/main/java/de/officeryoda/fhysics/engine/math/Vector2.kt
@@ -1,4 +1,4 @@
-package de.officeryoda.fhysics.engine
+package de.officeryoda.fhysics.engine.math
 
 import kotlin.math.cos
 import kotlin.math.sin
