+++ src/main/java/de/officeryoda/fhysics/engine/objects/Polygon.kt
@@ -1,12 +1,12 @@
 package de.officeryoda.fhysics.engine.objects
 
-import de.officeryoda.fhysics.engine.Matrix2x3
-import de.officeryoda.fhysics.engine.Projection
-import de.officeryoda.fhysics.engine.Vector2
 import de.officeryoda.fhysics.engine.collision.BorderEdge
 import de.officeryoda.fhysics.engine.collision.CollisionFinder
 import de.officeryoda.fhysics.engine.collision.CollisionInfo
 import de.officeryoda.fhysics.engine.collision.ContactFinder
+import de.officeryoda.fhysics.engine.math.Matrix2x3
+import de.officeryoda.fhysics.engine.math.Projection
+import de.officeryoda.fhysics.engine.math.Vector2
 import kotlin.math.abs
 import kotlin.math.cos
 import kotlin.math.sin
