+++ src/main/java/de/officeryoda/fhysics/engine/objects/PolygonCreator.kt
@@ -1,6 +1,6 @@
 package de.officeryoda.fhysics.engine.objects
 
-import de.officeryoda.fhysics.engine.Vector2
+import de.officeryoda.fhysics.engine.math.Vector2
 
 object PolygonCreator {
