+++ src/main/java/de/officeryoda/fhysics/engine/objects/Rectangle.kt
@@ -1,7 +1,7 @@
 package de.officeryoda.fhysics.engine.objects
 
-import de.officeryoda.fhysics.engine.Vector2
 import de.officeryoda.fhysics.engine.collision.CollisionInfo
+import de.officeryoda.fhysics.engine.math.Vector2
 import kotlin.math.cos
 import kotlin.math.sin
