+++ src/test/java/de/officeryoda/fhysics/engine/objects/RectangleTest.kt
@@ -1,6 +1,6 @@
 package de.officeryoda.fhysics.engine.objects
 
-import de.officeryoda.fhysics.engine.Vector2
+import de.officeryoda.fhysics.engine.math.Vector2
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
